Repo // Reorganize CandidateCellData contents.
